+++ lib/download-audio.js
@@ -1,6 +1,8 @@
 'use strict'
 
 const { EventEmitter } = require('events')
+const ytdl = require('ytdl-core')
+
 const horizon = require('horizon-youtube-mp3')
 const fue = require('file-utils-easy')
 const ytpl = require('ytpl')
@@ -140,26 +142,24 @@ class DownloadYTAudio extends EventEmitter {
     })
   }
 
-  getVideoInfo (videoId) {
-    return new Promise((resolve, reject) => {
-      horizon.getInfo(`${YOUTUBE_URL}${videoId}`, function (err, info) {
-        if (err) {
-          reject(new Error(err))
-          return
-        }
+  async getVideoInfo (videoId) {
+    const url = `${YOUTUBE_URL}${videoId}`
 
-        resolve({
-          id: videoId,
-          url: info.videoFile,
-          url_simple: `${YOUTUBE_URL}${videoId}`,
-          title: info.videoName,
-          thumbnail: info.videoThumb,
-          duration: info.videoTime,
-          author: null // TODO: need to improve/contribute to horizon lib
-        })
-      })
-    })
+    const advData = await ytdl.getBasicInfo(url)
+    return {
+      id: advData.video_id,
+      url: advData.video_url,
+      title: advData.title,
+      thumbnail: advData.player_response.videoDetails.thumbnail.thumbnails.reduce(getBigger),
+      duration: +(advData.player_response.videoDetails.lengthSeconds),
+      author: advData.author,
+      raw: advData
+    }
   }
 }
 
 module.exports = DownloadYTAudio
+
+function getBigger (a, b) {
+  return (a.width > b.width) ? a : b
+}
