+++ package.json
@@ -33,16 +33,19 @@
     "url": "https://github.com/Eomm/youtube-download-playlist/issues"
   },
   "dependencies": {
+    "commander": "^5.0.0",
     "file-utils-easy": "^1.2.0",
     "horizon-youtube-mp3": "^1.9.1",
     "rxjs": "^6.4.0",
+    "ytdl-core": "^2.0.0",
     "ytpl": "^0.1.14"
   },
   "devDependencies": {
     "@types/jest": "^23.3.14",
     "coveralls": "^3.0.3",
     "jest": "^22.4.4",
-    "standard": "^12.0.1"
+    "standard": "^12.0.1",
+    "tap": "^14.10.6"
   },
   "jest": {
     "testEnvironment": "node"
